+++ packages/smart-forms-renderer/src/components/FormComponents/ChoiceItems/ChoiceAutocompleteField.tsx
@@ -76,7 +76,7 @@ function ChoiceAutocompleteField(props: ChoiceAutocompleteFieldsProps) {
       value={valueCoding ?? null}
       options={options}
       getOptionLabel={(option) => option.display ?? `${option.code}`}
-      isOptionEqualToValue={(option, value) => option.id === value.id}
+      isOptionEqualToValue={(option, value) => option.code === value.code}
       disabled={readOnly && readOnlyVisualStyle === 'disabled'}
       readOnly={readOnly && readOnlyVisualStyle === 'readonly'}
       loading={loading}