+++ packages/smart-forms-renderer/src/components/FormComponents/OpenChoiceItems/OpenChoiceAutocompleteItem.tsx
@@ -72,10 +72,6 @@ function OpenChoiceAutocompleteItem(props: BaseItemProps) {
 
   const readOnly = useReadOnly(qItem, parentIsReadOnly);
 
-  // TODO Process calculated expressions
-  // This requires its own hook, because in the case of multi-select, we need to check if the value is already checked to prevent an infinite loop
-  // This will be done after the choice/open-choice refactoring
-
   // Perform validation checks
   const validationFeedback = useValidationFeedback(qItem, feedbackFromParent);
 
@@ -118,6 +114,7 @@ function OpenChoiceAutocompleteItem(props: BaseItemProps) {
     newValue: Coding | string | null,
     reason: AutocompleteChangeReason | string
   ) {
+    // When the reason is selectOption this function gets called twice. One for the string value and one for the selected option. So we need to make sure we ignore the string value on selectOption
     // If the reason is reset, then don't change value. Otherwise you will end up with looped setState calls
     if (reason === 'reset') {
       return;
@@ -133,7 +130,7 @@ function OpenChoiceAutocompleteItem(props: BaseItemProps) {
       newValue = '';
     }
 
-    if (typeof newValue === 'string') {
+    if (typeof newValue === 'string' && reason === 'input') {
       if (newValue !== '') {
         // Check if the newValue is in the options, first check options.display, then check options.code
         const foundOption = options.find((option) => {
@@ -159,8 +156,8 @@ function OpenChoiceAutocompleteItem(props: BaseItemProps) {
         // For empty string, update immediately
         onQrItemChange(createEmptyQrItem(qItem, answerKey));
       }
-    } else {
-      // For coding selection, update immediately (no debounce)
+    } else if (typeof newValue === 'object' && reason === 'selectOption') {
+      // For coding selection, update immediately (no debounce) i.e reason should be selectOption
       onQrItemChange({
         ...createEmptyQrItem(qItem, answerKey),
         answer: [{ id: answerKey, valueCoding: newValue }]